Add Note class
This class will allow for easy interfacing with notes in our database,
and is to be used by our team. The methods for interfacing with notes in
the database haven't been added to the class, and the functions for
creating Note objects to represent database notes have not been
implemented yet.