Prepare Note class for being a schema
Someone will need to fill out the two functions that are used to obtain
Note objects (Create(...) and FromDatabase(...)), as well as the method
that checks for the existence of a note in the database (Exists(...)).
If you end up doing that, use the Note constructor to prepare the Note
object (by filling it with field values from the database).